refactor: remove size screen duplicates
