--- src/lib/Components/ArtworkFilter/Filters/SizesOptions.tsx
@@ -1,45 +1,0 @@
-import { StackScreenProps } from "@react-navigation/stack"
-import { ArtworkFilterNavigationStack } from "lib/Components/ArtworkFilter"
-import { FilterData, FilterDisplayName, FilterParamName } from "lib/Components/ArtworkFilter/ArtworkFilterHelpers"
-import React from "react"
-import { MultiSelectOptionScreen } from "./MultiSelectOption"
-import { useMultiSelect } from "./useMultiSelect"
-
-interface SizesOptionsScreenProps extends StackScreenProps<ArtworkFilterNavigationStack, "SizesOptionsScreen"> {}
-
-export const SIZES_OPTIONS: FilterData[] = [
-  {
-    displayText: "Small (under 40cm)",
-    paramName: FilterParamName.sizes,
-    paramValue: "SMALL",
-  },
-  {
-    displayText: "Medium (40cm – 100cm)",
-    paramName: FilterParamName.sizes,
-    paramValue: "MEDIUM",
-  },
-  {
-    displayText: "Large (over 100cm)",
-    paramName: FilterParamName.sizes,
-    paramValue: "LARGE",
-  },
-]
-
-export const SizesOptionsScreen: React.FC<SizesOptionsScreenProps> = ({ navigation }) => {
-  const { handleSelect, isSelected, handleClear, isActive } = useMultiSelect({
-    options: SIZES_OPTIONS,
-    paramName: FilterParamName.sizes,
-  })
-
-  const filterOptions = SIZES_OPTIONS.map((option) => ({ ...option, paramValue: isSelected(option) }))
-
-  return (
-    <MultiSelectOptionScreen
-      onSelect={handleSelect}
-      filterHeaderText={FilterDisplayName.sizes}
-      filterOptions={filterOptions}
-      navigation={navigation}
-      {...(isActive ? { rightButtonText: "Clear", onRightButtonPress: handleClear } : {})}
-    />
-  )
-}
